requirements.txt
tensorflow-cpu==2.15.0
numpy==1.24.3
scipy==1.11.3
pandas==2.0.3
opencv-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.2
joblib==1.3.2
plotly==5.18.0
